Add files via upload
The project uses fingerprint to idenitfy the blood group of an individual.The system uses fingerprint similarities like ridges and whorls found in the fingers to identify the group. For the same, machine learning models have been used such as SIFT.These models compare the given samples with the ones we provide and give the result. The system has an accuracy of 60%.